# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GPSLink Gateway",
        "description": "Available API for GPSLink Gateway",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/user/list": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of registered users",
                "operationId": "UserList",
                "requestBody": {
                    "description": "Transporter Id - NOTE: If transporter_id object is omitted then all users will be return.",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "transporter_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "transporter_id": 0
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/userById/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by user id",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user id supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/update/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Updated user",
                "description": "Update user information.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "username_email": {
                                        "type": "string"
                                    },
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "transporter_id": {
                                        "type": "integer"
                                    },
                                    "contact_no": {
                                        "type": "string"
                                    },
                                    "user_role": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User is successfully updated!"
                    },
                    "400": {
                        "description": "Invalid user id supplied"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "409": {
                        "description": "User Id does not matched!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/updatePassword/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Updated user password",
                "description": "Update user password.",
                "operationId": "updatePassword",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Old and new password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "oldPassword": {
                                        "type": "string"
                                    },
                                    "newPassword": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Password is successfully updated!"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "449": {
                        "description": "New password is the same with the previous password used."
                    },
                    "409": {
                        "description": "Incorrect old password"
                    },
                    "500": {
                        "description": "Failed to update password!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/resetPassword/{id}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Reset user password",
                "description": "Reset user password.",
                "operationId": "ResetPassword",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id to be reset",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Password is successfully resetted!"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Failed to reset password!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs out current logged in user session",
                "description": "Logs out current logged in user session",
                "operationId": "logoutUser",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User",
                "required": [
                    "username_email",
                    "full_name",
                    "transporter_id",
                    "user_role"
                ],
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "username_email": {
                        "description": "Email",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Password",
                        "type": "string",
                        "format": "int64",
                        "maximum": 255
                    },
                    "full_name": {
                        "description": "Full name",
                        "type": "string"
                    },
                    "transporter_id": {
                        "description": "Transporter ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "contact_no": {
                        "description": "Phone",
                        "type": "string",
                        "format": "msisdn"
                    },
                    "user_role": {
                        "description": "User Role ID",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}